Fix code example in collectors post
The post was using an invalid code example. The text said one thing
but the code did something else. This was pointed out by a reader in
the comments.

The example about the combiner function was wrong and produces wrong
results when parallel streams are used. When merging maps, the values
should be summed, but previously, the code example only incremented one
value.